copy-tracking: move unit tests into backend specific file
